fix: don't include width and height if set to 0
This fix is related to an issue that usually happens with SVG. When an
SVG is saved, GraphCMS can't set its dimensions, so the width and height
of the SVG file are set to 0. So when rendering it, the image will be
hidden because the width/height is set to 0. This PR makes sure we only
add the width and height attributes if they are bigger than 0. Since the
SVG usually has a width and height defined in its document, it will be
rendered using its size.